Refactor the monitors out into separate files
